+++ aim/profiler/interfaces.py
@@ -1,11 +1,23 @@
+import os
+
 from aim.profiler.profiler import Profiler
 from aim.engine.utils import get_module
+from aim.engine.configs import AIM_PROFILER_ENABLED
 
 
 class BaseInterface:
     """
     Base interface for `Profiler` integration with various ML frameworks
     """
+
+    @staticmethod
+    def enabled():
+        """
+        Returns `True` if Profiler is enabled and collects statistics
+        or `False` otherwise
+        """
+        env = os.getenv(AIM_PROFILER_ENABLED)
+        return env and env == 'true'
 
     def label(self, *args, **kwargs):
         raise NotImplementedError('method not implemented')
@@ -21,13 +33,15 @@ class DefaultInterface(BaseInterface):
 
     @classmethod
     def label(cls, key):
-        p = Profiler()
-        p.track(key)
+        if cls.enabled():
+            p = Profiler()
+            p.label_tracking_start(key)
 
     @classmethod
     def loop(cls, key):
-        p = Profiler()
-        p.cycle(key)
+        if cls.enabled():
+            p = Profiler()
+            p.label_tracking_stop(key)
 
 
 class TensorFlowInterface(BaseInterface):
@@ -37,6 +51,7 @@ class TensorFlowInterface(BaseInterface):
 
     PROFILER_NODE_START = 'start'
     PROFILER_NODE_END = 'end'
+    PROFILER_NODE_CYCLE_END = 'cycle_end'
 
     tf = None
     profiler = None
@@ -45,16 +60,20 @@ class TensorFlowInterface(BaseInterface):
     def label(cls, key, inp):
         """
         Inserts node after `inp` node which will execute eagerly and
-        call profiler `label` function. That tells `Profiler` to start
-        tracking system statistics and collect them under `key` label
+        call profiler `label_tracking_start` function. That tells `Profiler`
+        to start tracking system statistics and collect them under `key` label.
         """
+        # Return input if profiler is not enabled
+        if not cls.enabled():
+            return inp
+
         cls._init()
         tf = cls.tf
 
         # Create TensorFlow op which wraps python function and calls eagerly
-        x = tf.py_function(
-            func=cls._profiler_node(key, cls.PROFILER_NODE_START),
-            inp=[inp], Tout=inp.dtype)
+        x = tf.stop_gradient(
+            tf.py_function(func=cls._profiler_node(cls.PROFILER_NODE_START,key),
+                           inp=[inp], Tout=inp.dtype))
 
         # Set node shape
         x.set_shape(inp.get_shape())
@@ -65,15 +84,42 @@ class TensorFlowInterface(BaseInterface):
     def loop(cls, key, inp):
         """
         Inserts node after `inp` node which will execute eagerly and
-        call profiler `loop` function. That tells `Profiler` to end
-        tracking system statistics under `key` label and aggregate
-        collected stats
+        call profiler `label_tracking_stop` function. That tells `Profiler`
+        to end tracking system statistics under `key` label and aggregate
+        collected stats if auto detection mode of cycles is enabled.
         """
+        # Return input if profiler is not enabled
+        if not cls.enabled():
+            return inp
+
         cls._init()
         tf = cls.tf
 
         # Create TensorFlow op which wraps python function and calls eagerly
-        x = tf.py_function(func=cls._profiler_node(key, cls.PROFILER_NODE_END),
+        x = tf.stop_gradient(
+            tf.py_function(func=cls._profiler_node(cls.PROFILER_NODE_END, key),
+                           inp=[inp], Tout=inp.dtype))
+
+        # Set node shape
+        x.set_shape(inp.get_shape())
+        return x
+
+    @classmethod
+    def cycle(cls, inp):
+        """
+        Inserts node after `inp` node which will execute eagerly and call
+        profiler `cycle_end` function, which will aggregate collected stats.
+        This operation can be useful if auto detection mode of is disabled.
+        """
+        # Return input if profiler is not enabled
+        if not cls.enabled():
+            return inp
+
+        cls._init()
+        tf = cls.tf
+
+        # Create TensorFlow op which wraps python function and calls eagerly
+        x = tf.py_function(func=cls._profiler_node(cls.PROFILER_NODE_CYCLE_END),
                            inp=[inp], Tout=inp.dtype)
 
         # Set node shape
@@ -92,33 +138,47 @@ class TensorFlowInterface(BaseInterface):
             cls.profiler = Profiler()
 
     @classmethod
-    def _profiler_node(cls, key, node_type):
+    def _profiler_node(cls, node_type, key=None):
         """
-        Returns python function to be called in `py_function` TensorFlow op
+        Returns python function that should be called within
+        `py_function` TensorFlow operation
         """
         def start(x):
             """
-            Calls `Profiler` `track` method, which indicates loop start
+            Calls `Profiler` `label_tracking_start` method,
+            which indicates loop start
             """
-            cls.profiler.track(key)
+            cls.profiler.label_tracking_start(key)
             return x
 
-        def end(x):
+        def stop(x):
             """
-            Calls `Profiler` `cycle` method, which indicates loop end
+            Calls `Profiler` `label_tracking_stop` method,
+            which indicates loop end
             """
-            cls.profiler.cycle(key)
+            cls.profiler.label_tracking_stop(key)
+            return x
+
+        def cycle(x):
+            """
+            Calls `Profiler` `cycle_end` method, which indicates a
+            full cycle was done
+            """
+            cls.profiler.cycle_end()
             return x
 
         if node_type == cls.PROFILER_NODE_START:
             return start
         elif node_type == cls.PROFILER_NODE_END:
-            return end
+            return stop
+        elif node_type == cls.PROFILER_NODE_CYCLE_END:
+            return cycle
 
 
 class KerasInterface(TensorFlowInterface):
     """
     Keras(only TF backend supported) interface for `Profiler` integration
+    TODO: Adjust to updated `Profiler` interface
     TODO: Add other backends support
     """
 
@@ -126,16 +186,23 @@ class KerasInterface(TensorFlowInterface):
 
     label_layer_cls = None
     loop_layer_cls = None
+    neutral_layer_cls = None
 
     @classmethod
     def label(cls, key, **kwargs):
+        if not cls.enabled():
+            return cls.neutral_layer_cls()
+
         cls._init()
         return cls.label_layer_cls(key)
 
     @classmethod
     def loop(cls, key, **kwargs):
+        if not cls.enabled():
+            return cls.neutral_layer_cls()
+
         cls._init()
-        return cls.label_layer_cls(key)
+        return cls.loop_layer_cls(key)
 
     @classmethod
     def _init(cls):
@@ -154,17 +221,17 @@ class KerasInterface(TensorFlowInterface):
 
         # Implement `keras` layer wrapper for profiler `track` method
         if cls.label_layer_cls is None:
-            class LabelLayer(cls.keras.layers.Layer):
+            class ProfilerLabelLayer(cls.keras.layers.Layer):
                 def __init__(self, key, **kwargs):
                     self.key = key
-                    super(LabelLayer, self).__init__(**kwargs)
+                    super(ProfilerLabelLayer, self).__init__(**kwargs)
 
                 def call(self, inp):
                     profiler_start_f = keras_interface.PROFILER_NODE_START
-                    x = tf.py_function(
+                    x = tf.stop_gradient(tf.py_function(
                         func=keras_interface._profiler_node(self.key,
                                                             profiler_start_f),
-                        inp=[inp], Tout=inp.dtype)
+                        inp=[inp], Tout=inp.dtype))
 
                     # Set node shape
                     x.set_shape(inp.get_shape())
@@ -173,21 +240,21 @@ class KerasInterface(TensorFlowInterface):
                 def compute_output_shape(self, input_shape):
                     return input_shape
 
-            cls.label_layer_cls = LabelLayer
+            cls.label_layer_cls = ProfilerLabelLayer
 
         # Implement `keras` layer wrapper for profiler `cycle` method
         if cls.loop_layer_cls is None:
-            class LoopLayer(cls.keras.layers.Layer):
+            class ProfilerLoopLayer(cls.keras.layers.Layer):
                 def __init__(self, key, **kwargs):
                     self.key = key
-                    super(LoopLayer, self).__init__(**kwargs)
+                    super(ProfilerLoopLayer, self).__init__(**kwargs)
 
                 def call(self, inp):
                     profiler_end_f = keras_interface.PROFILER_NODE_END
-                    x = tf.py_function(
+                    x = tf.stop_gradient(tf.py_function(
                         func=keras_interface._profiler_node(self.key,
                                                             profiler_end_f),
-                        inp=[inp], Tout=inp.dtype)
+                        inp=[inp], Tout=inp.dtype))
 
                     # Set node shape
                     x.set_shape(inp.get_shape())
@@ -196,4 +263,16 @@ class KerasInterface(TensorFlowInterface):
                 def compute_output_shape(self, input_shape):
                     return input_shape
 
-            cls.loop_layer_cls = LoopLayer
+            cls.loop_layer_cls = ProfilerLoopLayer
+
+        # Implement `keras` layer that does nothing, but passes input
+        # to the next layer. This layer is used when profiler is disabled.
+        if cls.neutral_layer_cls is None:
+            class ProfilerNeutralLayer(cls.keras.layers.Layer):
+                def call(self, inp):
+                    return tf.stop_gradient(inp)
+
+                def compute_output_shape(self, input_shape):
+                    return input_shape
+
+            cls.neutral_layer_cls = ProfilerNeutralLayer